man, this is hard but fun, been stuck for 2 days

+++ common_lib/src/message/message.rs
@@ -1,4 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+
+#[derive(Serialize, Deserialize, Debug)]
 pub struct JobCreationRequest {
     pub job_id : i64,
-    pub n: usize
+    pub n: i32
 }
